refactor: relocate artifact IO to ModelArtifactIO.swift (P4)
Weight-blob/sidecar loading, top-level weight-path resolvers, tokenizer
discovery, and metadata.json parsing move out of the engine (and out of
SplitHybridRuntime, where the P3 sweep parked loadConfigFromMetadataFile)
into Sources/RealModelInference/ModelArtifactIO.swift behind the frozen
engine extension. No behavior change.

Verified: golden replay bit-identical; full suite green.
