+++ Sources/RealModelInference/SplitHybridRuntime.swift
@@ -811,88 +811,6 @@ extension RealModelInferenceEngine {
         return (order, evictedKey)
     }
 
-    static func loadConfigFromMetadataFile(at metadataURL: URL) throws -> MultiModelConfig {
-        let data: Data
-        do {
-            data = try Data(contentsOf: metadataURL)
-        } catch {
-            throw RealModelInferenceError.runtimeFailure("Failed to read metadata.json: \(error)")
-        }
-
-        let object: Any
-        do {
-            object = try JSONSerialization.jsonObject(with: data)
-        } catch {
-            throw RealModelInferenceError.runtimeFailure("metadata.json is not valid JSON: \(error)")
-        }
-
-        guard let metadata = object as? [String: Any] else {
-            throw RealModelInferenceError.runtimeFailure("metadata.json must be a JSON object")
-        }
-
-        func requiredInt(_ key: String) throws -> Int {
-            guard let number = metadata[key] as? NSNumber else {
-                throw RealModelInferenceError.runtimeFailure("metadata.json missing numeric field \(key)")
-            }
-            return number.intValue
-        }
-
-        func requiredDouble(_ key: String) throws -> Double {
-            guard let number = metadata[key] as? NSNumber else {
-                throw RealModelInferenceError.runtimeFailure("metadata.json missing numeric field \(key)")
-            }
-            return number.doubleValue
-        }
-
-        guard let name = metadata["name"] as? String, !name.isEmpty else {
-            throw RealModelInferenceError.runtimeFailure("metadata.json missing string field name")
-        }
-        let architecture: MultiModelConfig.Architecture
-        switch (metadata["architecture"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() {
-        case "gpt2":
-            architecture = .gpt2
-        case "llama":
-            architecture = .llama
-        default:
-            throw RealModelInferenceError.runtimeFailure("metadata.json missing supported architecture")
-        }
-
-        let preferredDecodePath: MultiModelConfig.PreferredDecodePath?
-        if let value = metadata["preferredDecodePath"] {
-            let raw: String
-            if let string = value as? String {
-                raw = string
-            } else {
-                raw = String(describing: value)
-            }
-            do {
-                preferredDecodePath = try MultiModelConfig.PreferredDecodePath.parse(raw)
-            } catch {
-                throw RealModelInferenceError.runtimeFailure(
-                    "metadata.json has unsupported preferredDecodePath \"\(raw)\" (expected \"hybrid\" or \"exact_cpu\")"
-                )
-            }
-        } else {
-            preferredDecodePath = nil
-        }
-
-        return MultiModelConfig(
-            name: name,
-            nLayer: try requiredInt("nLayer"),
-            nHead: try requiredInt("nHead"),
-            nKVHead: try requiredInt("nKVHead"),
-            dModel: try requiredInt("dModel"),
-            headDim: try requiredInt("headDim"),
-            hiddenDim: try requiredInt("hiddenDim"),
-            vocab: try requiredInt("vocab"),
-            maxSeq: try requiredInt("maxSeq"),
-            normEps: Float(try requiredDouble("normEps")),
-            ropeTheta: Float((metadata["ropeTheta"] as? NSNumber)?.doubleValue ?? 10_000),
-            eosToken: (metadata["eosToken"] as? NSNumber)?.uint32Value,
-            architecture: architecture,
-            preferredDecodePath: preferredDecodePath
-        )
-    }
 
     static func resolveExactTwoTokenDraft(
         config: MultiModelConfig,
